ephemeral volumes and EIPs in CloudFormation; smarter cleanup

+++ modules/mu/cleanup.rb
@@ -30,6 +30,7 @@ module MU
     @deploy_id = nil
     @noop = false
     @onlycloud = false
+    @skipcloud = false
 
     # Purge all resources associated with a deployment.
     # @param deploy_id [String]: The identifier of the deployment to remove (typically seen in the MU-ID tag on a resource).
@@ -40,12 +41,18 @@ module MU
     # @param web [Boolean]: Generate web-friendly output.
     # @param ignoremaster [Boolean]: Ignore the tags indicating the originating MU master server when deleting.
     # @return [void]
-    def self.run(deploy_id, noop=false, skipsnapshots=false, onlycloud=false, verbose=false, web=false, ignoremaster=false, mommacat: nil)
+    def self.run(deploy_id, noop: false, skipsnapshots: false, onlycloud: false, verbose: false, web: false, ignoremaster: false, skipcloud: false, mommacat: nil)
       MU.setLogging(verbose, web)
       @noop = noop
       @skipsnapshots = skipsnapshots
       @onlycloud = onlycloud
+      @skipcloud = skipcloud
       @ignoremaster = ignoremaster
+
+      if @skipcloud and @onlycloud # you actually mean noop
+        @onlycloud = @skipcloud = false
+        @noop = true
+      end
 
       if MU.chef_user != "mu"
         MU.setVar("dataDir", Etc.getpwnam(MU.chef_user).dir+"/.mu/var")
@@ -75,46 +82,48 @@ module MU
         end
       end
 
-      parent_thread_id = Thread.current.object_id
-      regions = MU::Cloud::AWS.listRegions
-      deleted_nodes = 0
-      @regionthreads = []
-      keyname = "deploy-#{MU.deploy_id}"
-      regions.each { |r|
-        @regionthreads << Thread.new {
-          MU.dupGlobals(parent_thread_id)
-          MU.setVar("curRegion", r)
-          MU.log "Checking for cloud resources from #{MU.deploy_id} in #{r}", MU::NOTICE
+      if !@skipcloud
+        parent_thread_id = Thread.current.object_id
+        regions = MU::Cloud::AWS.listRegions
+        deleted_nodes = 0
+        @regionthreads = []
+        keyname = "deploy-#{MU.deploy_id}"
+        regions.each { |r|
+          @regionthreads << Thread.new {
+            MU.dupGlobals(parent_thread_id)
+            MU.setVar("curRegion", r)
+            MU.log "Checking for cloud resources from #{MU.deploy_id} in #{r}", MU::NOTICE
 
-          # We do these in an order that unrolls dependent resources sensibly,
-          # and we hit :Collection twice because AWS CloudFormation sometimes
-          # fails internally.
-          MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::ServerPool.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::LoadBalancer.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Server.cleanup(skipsnapshots: @skipsnapshots, onlycloud: @onlycloud, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Database.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::CacheCluster.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::FirewallRule.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Alarm.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Notification.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Log.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::DNSZone.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::VPC.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
-          MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r, wait: true)
+            # We do these in an order that unrolls dependent resources sensibly,
+            # and we hit :Collection twice because AWS CloudFormation sometimes
+            # fails internally.
+            MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::ServerPool.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::LoadBalancer.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Server.cleanup(skipsnapshots: @skipsnapshots, onlycloud: @onlycloud, noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Database.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::CacheCluster.cleanup(skipsnapshots: @skipsnapshots, noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::FirewallRule.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Alarm.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Notification.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Log.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::DNSZone.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::VPC.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r)
+            MU::Cloud::Collection.cleanup(noop: @noop, ignoremaster: @ignoremaster, region: r, wait: true)
 
-          resp = MU::Cloud::AWS.ec2(r).describe_key_pairs(
-              filters: [{name: "key-name", values: [keyname]}]
-          )
-          resp.data.key_pairs.each { |keypair|
-            MU.log "Deleting key pair #{keypair.key_name} from #{r}"
-            MU::Cloud::AWS.ec2(r).delete_key_pair(key_name: keypair.key_name) if !@noop
+            resp = MU::Cloud::AWS.ec2(r).describe_key_pairs(
+                filters: [{name: "key-name", values: [keyname]}]
+            )
+            resp.data.key_pairs.each { |keypair|
+              MU.log "Deleting key pair #{keypair.key_name} from #{r}"
+              MU::Cloud::AWS.ec2(r).delete_key_pair(key_name: keypair.key_name) if !@noop
+            }
           }
         }
-      }
 
-      @regionthreads.each do |t|
-        t.join
+        @regionthreads.each do |t|
+          t.join
+        end
       end
 
       # Scrub any residual Chef records with matching tags
@@ -226,14 +235,14 @@ module MU
         end
       end
 
-      if !@noop
+      if !@noop and !@skipcloud
         MU::Cloud::AWS.s3(MU.myRegion).delete_object(
             bucket: MU.adminBucketName,
             key: "#{MU.deploy_id}-secret"
         )
+        MU::Cloud::AWS.openFirewallForClients # XXX should only run if we're in AWS...
       end
 
-      MU::Cloud::AWS.openFirewallForClients # XXX should only run if we're in AWS...
       MU::MommaCat.syncMonitoringConfig if !@noop
 
     end
